requirements-dev.txt
appdirs==1.4.4
black==21.5b1
click==8.0.0
jeepney==0.6.0
loguru==0.5.3
mypy-extensions==0.4.3
notify-py==0.3.2
pathspec==0.8.1
PyYAML==5.4.1
regex==2021.4.4
toml==0.10.2
